change style, misc renaming

+++ Adria/Utilities/AllocatorUtil.h
@@ -3,14 +3,14 @@
 namespace adria
 {
 
-	inline size_t Align(size_t address, size_t align)
+	inline Uint64 Align(Uint64 address, Uint64 align)
 	{
 		if ((0 == align) || (align & (align - 1))) return address;
 
 		return ((address + (align - 1)) & ~(align - 1));
 	}
 
-	using OffsetType = size_t;
+	using OffsetType = Uint64;
 
 	inline constexpr OffsetType const INVALID_OFFSET = static_cast<OffsetType>(-1);
